Add sampling year coloring
For builds where we've chosen not to infer a timetree it's still very
useful to communicate the sampling date. Here we export the year which
forms a nice continuous coloring. (An ordinal coloring would be better
but the date range in this dataset means Auspice repeats colors which
isn't ideal. Temporal coloring will be better once it accepts YYYY-MM-DD
strings, but as of Auspice 2.55 it interprets integers as January 1st.)